Moved functionality from CodeEditor to base class LineNumberEditor
