+++ super32emu/super32emu/ui/code_editor.py
@@ -1,9 +1,8 @@
-from .line_number_editor import LineNumberEditor
-from PySide2.QtCore import SIGNAL, QRect, QObject
-from PySide2.QtGui import QColor, Qt, QTextFormat, QPainter, QMouseEvent
 from PySide2.QtCore import SIGNAL, QRect
-from PySide2.QtGui import QColor, Qt, QTextFormat, QPainter, QTextCursor
-from PySide2.QtWidgets import QPlainTextEdit, QTextEdit
+from PySide2.QtGui import QMouseEvent
+from PySide2.QtGui import Qt, QPainter
+
+from .line_number_editor import LineNumberEditor
 
 
 class CodeEditor(LineNumberEditor):
@@ -18,25 +17,8 @@ class CodeEditor(LineNumberEditor):
         self.lineNumberArea.mouseReleaseEvent = self.onClicked
         self.breakpoints = []
 
-        self.connect(self, SIGNAL('blockCountChanged(int)'), self.updateLineNumberAreaWidth)
-        self.connect(self, SIGNAL('updateRequest(QRect,int)'), self.updateLineNumberArea)
-        self.connect(self, SIGNAL('cursorPositionChanged()'), self.highlightCurrentLine)
-
         # It is necessary to calculate the line number area width when the editor is created
         self.updateLineNumberAreaWidth(0)
-
-    def lineNumberAreaWidth(self):
-        """
-        The lineNumberAreaWidth() function calculates the width of the LineNumberArea widget.
-        We take the number of digits in the last line of the editor and multiply that with the maximum width of a digit.
-        """
-        digits = 1
-        count = max(1, self.blockCount())
-        while count >= 10:
-            count /= 10
-            digits += 1
-        space = 3 + self.fontMetrics().width('9') * digits
-        return space
 
     def updateLineNumberAreaWidth(self, _):
         self.setViewportMargins(self.lineNumberAreaWidth(), 0, 0, 0)
@@ -112,18 +94,3 @@ class CodeEditor(LineNumberEditor):
             block = block.next()
             top = bottom
             bottom = top + self.blockBoundingRect(block).height()
-
-    def resetHighlightedLines(self):
-        self.extraSelections = []
-        self.setExtraSelections(self.extraSelections)
-
-    def highlightLine(self, line_number: int, color=Qt.yellow):
-        lineColor = QColor(color)
-
-        selection = QTextEdit.ExtraSelection()
-        selection.format.setBackground(lineColor)
-        selection.format.setProperty(QTextFormat.FullWidthSelection, True)
-        selection.cursor = QTextCursor(self.document().findBlockByLineNumber(line_number))
-        selection.cursor.clearSelection()
-        self.extraSelections.append(selection)
-        self.setExtraSelections(self.extraSelections)
